// frontend/src/api/generated/models/TemplateUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Обновление шаблона (partial update).
 */
export type TemplateUpdate = {
    name?: (string | null);
    channel?: (string | null);
    subject_template?: (string | null);
    body_template?: (string | null);
    is_active?: (boolean | null);
};